Support cleaning (scrubbing) of genes
scrub_gene() will now go through each allele in a gene, checking
if it contributes positively to the final image. If not, the allele
is removed from the gene.

Currently, the end result is 0.75% worse than the original, but at
the benefit of around 120 alleles per thousand alleles in a gene.

This will speed future calculations, so all in all it's a net positive.

Why it gets worse you ask? I'm not quite sure right now, but it
might have to do with partially obscured circles. Not sure.