added two new mutations
color_mutation
Take an existing object, and change its color slightly. Insert it next
to the original object.

sideways_mutation
Take an existing object, and make a copy. The copy's location is
shifted a bit to the side and made a bit smaller. Insert it next
to the original object.